Break apart components, start styling
