+++ packages/web/src/assets/styles/index.css
@@ -3,6 +3,7 @@
 @import '~assets/styles/type.css';
 @import '~assets/styles/emoji.css';
 @import '~assets/styles/ant-overrides.css';
+@import '~assets/styles/shadows.css';
 
 #root *,
 body * {
