+++ packages/web/src/pages/chat-page/ChatPage.module.css
@@ -22,7 +22,7 @@
 }
 
 .chatList {
-  width: 256px;
+  width: 400px;
 }
 
 .messages {
