+++ packages/web/src/pages/chat-page/components/ChatMessageList.module.css
@@ -2,28 +2,7 @@
   display: flex;
   flex-direction: column-reverse;
   gap: 16px;
-  padding: 8px;
+  padding: 32px;
   overflow-y: auto;
-  background-color: var(--white);
-}
-
-.message {
-  display: flex;
-  gap: 8px;
-}
-
-.message.isAuthor {
-  flex-direction: row-reverse;
-}
-
-.messageBubble {
-  align-self: center;
-  background-color: var(--neutral-light-8);
-  padding: 12px;
-  border-radius: 16px;
-}
-
-.messageBubbleText {
-  word-break: break-word;
-  text-align: left;
+  background-color: var(--background);
 }
