+++ packages/web/src/pages/chat-page/components/ChatMessageList.tsx
@@ -1,49 +1,16 @@
 import { ComponentPropsWithoutRef, useEffect } from 'react'
 
-import {
-  accountSelectors,
-  chatActions,
-  chatSelectors,
-  decodeHashId,
-  cacheUsersSelectors
-} from '@audius/common'
-import type { ChatMessage } from '@audius/sdk'
+import { chatActions, chatSelectors } from '@audius/common'
 import cn from 'classnames'
 import { useDispatch } from 'react-redux'
 
 import { useSelector } from 'common/hooks/useSelector'
-import { ProfilePicture } from 'components/notification/Notification/components/ProfilePicture'
 
 import styles from './ChatMessageList.module.css'
+import { ChatMessageListItem } from './ChatMessageListItem'
 
 const { fetchNewChatMessages } = chatActions
 const { getChatMessages } = chatSelectors
-
-type ChatMessageListItemProps = {
-  message: ChatMessage
-}
-
-const ChatMessageListItem = (props: ChatMessageListItemProps) => {
-  const { message } = props
-  const senderUserId = decodeHashId(message.sender_user_id)
-  const userId = useSelector(accountSelectors.getUserId)
-  const isAuthor = userId === senderUserId
-  const user = useSelector((state) =>
-    cacheUsersSelectors.getUser(state, { id: senderUserId })
-  )
-  return (
-    <div
-      className={cn(styles.message, {
-        [styles.isAuthor]: isAuthor
-      })}
-    >
-      {user ? <ProfilePicture user={user} /> : null}
-      <div className={styles.messageBubble}>
-        <div className={styles.messageBubbleText}>{message.message}</div>
-      </div>
-    </div>
-  )
-}
 
 type ChatMessageListProps = ComponentPropsWithoutRef<'div'> & {
   chatId?: string
@@ -59,7 +26,7 @@ export const ChatMessageList = (props: ChatMessageListProps) => {
       if (chatId) {
         dispatch(fetchNewChatMessages({ chatId }))
       }
-    }, 1000)
+    }, 10000)
     return () => {
       clearInterval(pollInterval)
     }
